perf(plugin): optimize aws s3 file listing
Lower number of requests sent to S3 while listing objects. For buckets
with large number of objects, listing can take some time, because there
is a request for metadata sent to s3 for each object in the bucket.
This is redundant, because all data needed for objects listing are
available in S3ObjectSummary. This reduce listing time significantly.

Resolves: #490